fix(BCHEP-672): Replace ZipfileJob/PdfGenerationJob with GeneratePdfJob, a proper Sidekiq worker with deduplication lock and 30s popen3 timeout with process group kill to prevent thread starvation. Also fixes raw enum display values in PDF renderer and removes ~800 lines of dead debug code.
