--- app/jobs/pdf_generation_job.rb
@@ -1,857 +1,0 @@
-require "open3"
-require "json"
-require "fileutils"
-
-class PdfGenerationJob
-  include Sidekiq::Worker
-  sidekiq_options lock: :until_and_while_executing,
-                  queue: :file_processing,
-                  on_conflict: {
-                    client: :log,
-                    server: :reject
-                  }
-
-  def self.lock_args(args)
-    ## only lock on the first argument, which is the permit application id
-    ## this will prevent multiple jobs from running for the same permit application
-    [args[0]]
-  end
-
-  def perform(permit_application_id)
-    permit_application = PermitApplication.find(permit_application_id)
-    return if permit_application.blank?
-
-    generation_directory_path = Rails.root.join("tmp/files")
-    asset_directory_path = Rails.root.join("public")
-
-    puts "=== Directory Creation Debug ==="
-    puts "Target directory: #{generation_directory_path}"
-    puts "Rails root: #{Rails.root}"
-    puts "Directory exists before creation: #{File.directory?(generation_directory_path)}"
-    puts "Parent directory (/app/tmp) exists: #{File.directory?(Rails.root.join("tmp"))}"
-    puts "Parent directory permissions: #{File.directory?(Rails.root.join("tmp")) ? File.stat(Rails.root.join("tmp")).mode.to_s(8) : "N/A"}"
-
-    # Ensure the directory exists and has correct permissions
-    unless File.directory?(generation_directory_path)
-      puts "Creating directory: #{generation_directory_path}"
-      begin
-        FileUtils.mkdir_p(generation_directory_path)
-        puts "mkdir_p completed without error"
-      rescue => e
-        puts "ERROR during mkdir_p: #{e.class}: #{e.message}"
-        raise e
-      end
-    end
-
-    puts "Directory exists after creation attempt: #{File.directory?(generation_directory_path)}"
-
-    # Ensure directory is writable
-    if File.directory?(generation_directory_path)
-      unless File.writable?(generation_directory_path)
-        puts "Directory not writable, setting permissions"
-        FileUtils.chmod(0755, generation_directory_path)
-      end
-      puts "Directory is writable: #{File.writable?(generation_directory_path)}"
-      puts "Directory permissions: #{File.stat(generation_directory_path).mode.to_s(8)}"
-    else
-      puts "ERROR: Directory still does not exist after creation attempt!"
-      raise "Failed to create directory: #{generation_directory_path}"
-    end
-
-    puts "Directory ensured: #{generation_directory_path}"
-    puts "=== End Directory Debug ==="
-
-    submission_version_with_missing_pdfs =
-      permit_application.submission_versions.select(&:missing_pdfs?)
-
-    submission_versions_data =
-      submission_version_with_missing_pdfs.map do |submission_version|
-        # Convert data to JSON string
-        application_filename =
-          "permit_application_#{permit_application.id}_v#{submission_version.version_number}.pdf"
-        step_code_filename =
-          "step_code_checklist_#{permit_application.id}_v#{submission_version.version_number}.pdf"
-
-        should_permit_application_pdf_be_generated =
-          submission_version.missing_permit_application_pdf?
-        should_checklist_pdf_be_generated =
-          submission_version.missing_step_code_checklist_pdf?
-
-        permit_app_data =
-          camelize_response(
-            PermitApplicationBlueprint.render_as_json(
-              permit_application,
-              view: :pdf_generation,
-              form_json: submission_version.form_json,
-              submission_data: submission_version.formatted_submission_data,
-              submitted_at: submission_version.created_at
-            )
-          )
-
-        puts "=== PDF Data Generation Debug ==="
-        puts "Permit application data keys: #{permit_app_data&.keys&.join(", ")}"
-        puts "Permit application ID: #{permit_app_data&.dig("id")}"
-        puts "Permit application number: #{permit_app_data&.dig("number")}"
-        puts "Submitter data: #{permit_app_data&.dig("submitter")&.inspect}"
-        puts "Has step code checklist: #{submission_version.has_step_code_checklist?}"
-        puts "Should generate permit app PDF: #{should_permit_application_pdf_be_generated}"
-        puts "Should generate checklist PDF: #{should_checklist_pdf_be_generated}"
-
-        pdf_json_data = {
-          permitApplication: permit_app_data,
-          checklist:
-            submission_version.has_step_code_checklist? &&
-              camelize_response(submission_version.step_code_checklist_json),
-          meta: {
-            generationPaths: {
-              permitApplication:
-                should_permit_application_pdf_be_generated &&
-                  generation_directory_path.join(application_filename).to_s,
-              stepCodeChecklist:
-                should_checklist_pdf_be_generated &&
-                  generation_directory_path.join(step_code_filename).to_s
-            },
-            assetDirectoryPath: asset_directory_path.to_s
-          }
-        }.to_json
-
-        puts "Meta generationPaths: #{JSON.parse(pdf_json_data)["meta"]["generationPaths"].inspect}"
-
-        # Additional JSON data debugging
-        parsed_json = JSON.parse(pdf_json_data)
-        puts "JSON structure keys: #{parsed_json.keys.join(", ")}"
-        puts "Permit application keys: #{parsed_json["permitApplication"]&.keys&.join(", ")}"
-        puts "Form customizations present: #{parsed_json["permitApplication"]&.dig("formCustomizations").present?}"
-        puts "Form customizations content: #{parsed_json["permitApplication"]&.dig("formCustomizations")&.inspect}"
-        puts "Form JSON present: #{parsed_json["permitApplication"]&.dig("formJson").present?}"
-        puts "Submission data present: #{parsed_json["permitApplication"]&.dig("submissionData").present?}"
-        puts "=== End PDF Data Debug ==="
-
-        {
-          submission_version: submission_version,
-          pdf_json_data: pdf_json_data,
-          application_filename: application_filename,
-          step_code_filename: step_code_filename,
-          should_permit_application_pdf_be_generated:
-            should_permit_application_pdf_be_generated,
-          should_checklist_pdf_be_generated: should_checklist_pdf_be_generated
-        }
-      end
-
-    submission_versions_data.each do |submission_version_data|
-      generate_pdfs(submission_version_data, generation_directory_path)
-    end
-  end
-
-  def generate_pdfs(submission_version_data, generation_directory_path)
-    submission_version = submission_version_data[:submission_version]
-    pdf_json_data = submission_version_data[:pdf_json_data]
-    application_filename = submission_version_data[:application_filename]
-    step_code_filename = submission_version_data[:step_code_filename]
-    should_permit_application_pdf_be_generated =
-      submission_version_data[:should_permit_application_pdf_be_generated]
-    should_checklist_pdf_be_generated =
-      submission_version_data[:should_checklist_pdf_be_generated]
-
-    json_filename =
-      "#{generation_directory_path}/pdf_json_data_#{submission_version.id}.json"
-
-    File.open(json_filename, "w") { |file| file.write(pdf_json_data) }
-
-    # Verify JSON file was written correctly
-    puts "=== JSON File Verification ==="
-    if File.exist?(json_filename)
-      json_content = File.read(json_filename)
-      puts "JSON file written successfully"
-      puts "JSON file readable: #{!json_content.empty?}"
-
-      begin
-        parsed_verification = JSON.parse(json_content)
-        puts "JSON is valid and parseable"
-        puts "Parsed keys: #{parsed_verification.keys.join(", ")}"
-
-        # Check critical data points
-        permit_app = parsed_verification["permitApplication"]
-        if permit_app
-          puts "Permit application data present: ✓"
-          puts "Permit app ID: #{permit_app["id"]}"
-          puts "Form JSON present: #{permit_app["formJson"].present?}"
-          puts "Submission data present: #{permit_app["submissionData"].present?}"
-        else
-          puts "ERROR: No permit application data in JSON!"
-        end
-
-        meta = parsed_verification["meta"]
-        if meta && meta["generationPaths"]
-          puts "Generation paths present: ✓"
-          puts "Permit app path: #{meta["generationPaths"]["permitApplication"]}"
-        else
-          puts "ERROR: No generation paths in JSON!"
-        end
-      rescue JSON::ParserError => e
-        puts "ERROR: JSON is not valid - #{e.message}"
-      end
-    else
-      puts "ERROR: JSON file was not created!"
-    end
-    puts "=== End JSON Verification ==="
-
-    puts "=== Node.js Execution Debug ==="
-    puts "JSON filename: #{json_filename}"
-    puts "JSON file exists: #{File.exist?(json_filename)}"
-    puts "JSON file size: #{File.exist?(json_filename) ? File.size(json_filename) : "N/A"} bytes"
-    puts "Working directory: #{Rails.root}"
-    puts "Command: npm run #{NodeScripts::GENERATE_PDF_SCRIPT_NAME} #{json_filename}"
-    puts "Expected output files:"
-    puts "- Permit application PDF: #{should_permit_application_pdf_be_generated ? "#{generation_directory_path}/#{application_filename}" : "Not generating"}"
-    puts "- Step code checklist PDF: #{should_checklist_pdf_be_generated ? "#{generation_directory_path}/#{step_code_filename}" : "Not generating"}"
-    puts "JSON content preview (first 500 chars): #{pdf_json_data[0..500]}..."
-    puts "Files in directory before execution: #{Dir.entries(generation_directory_path).join(", ")}"
-
-    # Debug the exact paths that will be generated
-    parsed_json = JSON.parse(pdf_json_data)
-    generation_paths = parsed_json.dig("meta", "generationPaths")
-    if generation_paths
-      puts "=== Expected Output Paths ==="
-      generation_paths.each do |key, path|
-        next unless path
-        puts "#{key}: #{path}"
-        puts "  - Directory exists: #{File.directory?(File.dirname(path))}"
-        puts "  - Directory writable: #{File.writable?(File.dirname(path))}"
-        puts "  - Path absolute: #{File.absolute_path?(path)}"
-      end
-      puts "=== End Expected Paths ==="
-    end
-
-    # Check Node.js environment
-    puts "=== Node.js Environment Check ==="
-    node_version =
-      begin
-        `node --version 2>&1`.strip
-      rescue StandardError
-        "Not available"
-      end
-    npm_version =
-      begin
-        `npm --version 2>&1`.strip
-      rescue StandardError
-        "Not available"
-      end
-    puts "Node.js version: #{node_version}"
-    puts "NPM version: #{npm_version}"
-    puts "Package.json exists: #{File.exist?(Rails.root.join("package.json"))}"
-    puts "SSR script exists: #{File.exist?(Rails.root.join("public/vite-ssr/ssr.js"))}"
-    puts "=== End Environment Check ==="
-
-    # Execute Node.js PDF generation
-    puts "=== Executing Node.js PDF Generation ==="
-    command =
-      "npm run #{NodeScripts::GENERATE_PDF_SCRIPT_NAME} #{json_filename}"
-    puts "Running command: #{command}"
-
-    # Check files before Node.js execution
-    expected_files = []
-    if should_permit_application_pdf_be_generated
-      expected_files << "#{generation_directory_path}/#{application_filename}"
-    end
-    if should_checklist_pdf_be_generated
-      expected_files << "#{generation_directory_path}/#{step_code_filename}"
-    end
-
-    puts "Expected files to be created: #{expected_files.join(", ")}"
-
-    # Capture Node.js execution with enhanced logging
-    start_time = Time.current
-    stdout, stderr, status = Open3.capture3(command, chdir: Rails.root)
-    execution_time = Time.current - start_time
-
-    puts "=== Node.js Command Output ==="
-    puts "Exit status: #{status.exitstatus}"
-    puts "Execution time: #{execution_time.round(2)}s"
-    puts "Command PID: #{status.pid}" if status.respond_to?(:pid)
-
-    # Always log stdout/stderr, even if empty
-    puts "STDOUT (#{stdout.length} chars):"
-    if stdout.blank?
-      puts "  [EMPTY - This might indicate silent failure]"
-    else
-      puts "  #{stdout}"
-    end
-
-    puts "STDERR (#{stderr.length} chars):"
-    if stderr.blank?
-      puts "  [EMPTY]"
-    else
-      puts "  #{stderr}"
-    end
-
-    # Check for suspicious patterns
-    suspicious_patterns = []
-    if stdout.blank? && stderr.blank?
-      suspicious_patterns << "No output from Node.js script (completely silent)"
-    end
-    if execution_time < 0.5
-      suspicious_patterns << "Execution too fast (#{execution_time.round(2)}s) - likely early exit"
-    end
-    if execution_time > 30
-      suspicious_patterns << "Execution too slow (#{execution_time.round(2)}s) - possible timeout"
-    end
-
-    if suspicious_patterns.any?
-      puts "⚠️  SUSPICIOUS PATTERNS DETECTED:"
-      suspicious_patterns.each { |pattern| puts "   - #{pattern}" }
-    end
-
-    puts "=== End Node.js Output ==="
-
-    unless status.success?
-      puts "ERROR: Node.js PDF generation failed with exit code #{status.exitstatus}"
-      puts "STDOUT: #{stdout}"
-      puts "STDERR: #{stderr}"
-      raise "Node.js PDF generation failed: #{stderr.blank? ? stdout : stderr}"
-    end
-
-    # Immediately check if expected files were created
-    puts "=== Post-Node.js File Verification ==="
-    expected_files.each do |expected_file|
-      if File.exist?(expected_file)
-        file_size = File.size(expected_file)
-        puts "✓ File created: #{expected_file} (#{file_size} bytes)"
-
-        # Check if file is suspiciously small
-        if file_size < 10_000 # Less than 10KB might indicate issues
-          puts "⚠️  WARNING: File is smaller than expected (#{file_size} bytes)"
-          puts "   This might indicate React PDF rendering issues"
-        end
-      else
-        puts "✗ MISSING: #{expected_file}"
-        puts "   Node.js exited successfully but didn't create this file"
-
-        # Check for common issues
-        dir_exists = File.directory?(File.dirname(expected_file))
-        dir_writable = dir_exists && File.writable?(File.dirname(expected_file))
-        puts "   Directory exists: #{dir_exists}"
-        puts "   Directory writable: #{dir_writable}" if dir_exists
-
-        # Log a more detailed error
-        if stdout.include?("PDF Generation Error:") || stderr.include?("Error:")
-          puts "   Detected error messages in Node.js output"
-        else
-          puts "   No error messages in Node.js output - likely React PDF silent failure"
-        end
-      end
-    end
-    puts "=== End File Verification ==="
-
-    puts "Node.js PDF generation completed with exit status 0"
-
-    # Collect generated PDF files
-    pdfs = []
-    if should_permit_application_pdf_be_generated
-      pdfs << {
-        fname: application_filename,
-        key: PermitApplication::PERMIT_APP_PDF_DATA_KEY
-      }
-    end
-
-    if should_checklist_pdf_be_generated
-      pdfs << {
-        fname: step_code_filename,
-        key: PermitApplication::CHECKLIST_PDF_DATA_KEY
-      }
-    end
-
-    # Wait a moment for file system to sync (prevents timing issues)
-    sleep(0.5)
-
-    # Process the generated PDFs
-    pdfs.each do |pdf|
-      path = "#{generation_directory_path}/#{pdf[:fname]}"
-      puts "Processing PDF: #{path}"
-      puts "File exists: #{File.exist?(path)}"
-      puts "File size: #{File.exist?(path) ? File.size(path) : "N/A"} bytes"
-
-      # Check if file exists with retry for timing issues
-      retries = 0
-      max_retries = 3
-      while !File.exist?(path) && retries < max_retries
-        puts "PDF file not found, waiting and retrying... (attempt #{retries + 1}/#{max_retries})"
-        sleep(1)
-        retries += 1
-      end
-
-      unless File.exist?(path)
-        puts "PDF file was not created by Node.js script: #{path}"
-        puts "=== Diagnosing Node.js PDF Generation Failure ==="
-
-        # Comprehensive failure analysis - we ALWAYS find a reason
-        failure_reasons = []
-        confidence_scores = {}
-
-        # 1. Analyze Node.js execution patterns
-        puts "Analyzing Node.js execution patterns..."
-        if defined?(execution_time) && execution_time
-          if execution_time < 0.5
-            failure_reasons << "Node.js script exited too quickly (#{execution_time.round(2)}s) - likely early termination"
-            confidence_scores["early_exit"] = 90
-          elsif execution_time > 30
-            failure_reasons << "Node.js script took too long (#{execution_time.round(2)}s) - likely timeout or hang"
-            confidence_scores["timeout"] = 85
-          end
-        end
-
-        if defined?(stdout) && defined?(stderr)
-          if stdout.blank? && stderr.blank?
-            failure_reasons << "Node.js produced no output (completely silent execution)"
-            confidence_scores["silent_execution"] = 80
-          end
-        end
-
-        # 2. Check JSON data integrity with detailed analysis
-        puts "Analyzing JSON data integrity..."
-        if File.exist?(json_filename)
-          begin
-            json_content = JSON.parse(File.read(json_filename))
-            json_size = File.size(json_filename)
-
-            # Deep data validation
-            if json_content.dig("permitApplication", "formJson").blank?
-              failure_reasons << "Critical data missing: formJson is empty or null"
-              confidence_scores["missing_form_json"] = 95
-            else
-              form_json = json_content.dig("permitApplication", "formJson")
-              if form_json.is_a?(Hash) && form_json.dig("components").blank?
-                failure_reasons << "FormJSON has no components - invalid form structure"
-                confidence_scores["invalid_form_structure"] = 90
-              end
-            end
-
-            if json_content.dig("permitApplication", "submissionData").blank?
-              failure_reasons << "Critical data missing: submissionData is empty"
-              confidence_scores["missing_submission_data"] = 85
-            end
-
-            if json_content.dig(
-                 "meta",
-                 "generationPaths",
-                 "permitApplication"
-               ).blank?
-              failure_reasons << "Missing generation path configuration"
-              confidence_scores["missing_paths"] = 95
-            end
-
-            # Check for data corruption
-            if json_size > 100_000 # Very large JSON > 100KB
-              failure_reasons << "JSON data unusually large (#{json_size} bytes) - possible data corruption or circular references"
-              confidence_scores["large_data"] = 70
-            end
-          rescue JSON::ParserError => e
-            failure_reasons << "Invalid JSON data format: #{e.message}"
-            confidence_scores["json_parse_error"] = 95
-          end
-        else
-          failure_reasons << "JSON input file missing or deleted"
-          confidence_scores["missing_json"] = 95
-        end
-
-        # 3. System resource analysis
-        puts "Checking system resources..."
-        begin
-          # Memory check
-          memory_info =
-            `cat /proc/meminfo | grep MemAvailable | awk '{print $2}'`.strip.to_i
-          if memory_info > 0
-            if memory_info < 200_000 # Less than 200MB
-              failure_reasons << "Critically low memory (#{(memory_info / 1024).round}MB available) - likely out of memory during PDF rendering"
-              confidence_scores["low_memory"] = 95
-            elsif memory_info < 500_000 # Less than 500MB
-              failure_reasons << "Low memory available (#{(memory_info / 1024).round}MB) - may cause React PDF failures"
-              confidence_scores["moderate_memory"] = 70
-            end
-          end
-
-          # Disk space check
-          disk_info =
-            `df #{generation_directory_path} | tail -1 | awk '{print $4}'`.strip.to_i
-          if disk_info > 0
-            if disk_info < 50_000 # Less than 50MB
-              failure_reasons << "Critically low disk space (#{(disk_info / 1024).round}MB available) - cannot write PDF file"
-              confidence_scores["low_disk"] = 95
-            elsif disk_info < 100_000 # Less than 100MB
-              failure_reasons << "Low disk space (#{(disk_info / 1024).round}MB available) - may prevent PDF creation"
-              confidence_scores["moderate_disk"] = 70
-            end
-          end
-        rescue => e
-          failure_reasons << "Unable to check system resources: #{e.message}"
-          confidence_scores["resource_check_failed"] = 60
-        end
-
-        # 4. React PDF specific analysis
-        puts "Analyzing React PDF rendering issues..."
-        if File.exist?(json_filename)
-          begin
-            json_content = JSON.parse(File.read(json_filename))
-            permit_data = json_content.dig("permitApplication")
-
-            # Check for data that commonly breaks React PDF
-            if permit_data
-              # Check for potential circular references
-              data_string = permit_data.to_s
-              if data_string.length > 500_000 # Very large serialized data
-                failure_reasons << "Permit application data extremely large - likely contains circular references or excessive nested data"
-                confidence_scores["circular_references"] = 80
-              end
-
-              # Check for problematic field types
-              if permit_data.to_s.include?("null") &&
-                   permit_data.to_s.scan(/null/).count > 100
-                failure_reasons << "Excessive null values in data - may cause React component rendering issues"
-                confidence_scores["excessive_nulls"] = 75
-              end
-
-              # Check form complexity
-              form_json = permit_data.dig("formJson")
-              if form_json.is_a?(Hash) &&
-                   form_json.dig("components").is_a?(Array)
-                component_count =
-                  count_nested_components(form_json.dig("components"))
-                if component_count > 500
-                  failure_reasons << "Form has #{component_count} components - exceeds React PDF rendering limits"
-                  confidence_scores["complex_form"] = 85
-                end
-              end
-            end
-          rescue => e
-            failure_reasons << "Error analyzing React PDF data: #{e.message}"
-            confidence_scores["react_analysis_failed"] = 60
-          end
-        end
-
-        # 5. Environment and dependency issues
-        puts "Checking Node.js environment..."
-        begin
-          node_modules_exists = File.directory?(Rails.root.join("node_modules"))
-          if !node_modules_exists
-            failure_reasons << "Node modules directory missing - dependencies not installed"
-            confidence_scores["missing_dependencies"] = 95
-          end
-
-          react_pdf_exists =
-            File.directory?(Rails.root.join("node_modules", "@react-pdf"))
-          if !react_pdf_exists
-            failure_reasons << "@react-pdf/renderer package missing or not installed"
-            confidence_scores["missing_react_pdf"] = 95
-          end
-        rescue => e
-          failure_reasons << "Error checking Node.js environment: #{e.message}"
-          confidence_scores["env_check_failed"] = 60
-        end
-
-        # 6. If no specific reasons found, provide detailed generic analysis
-        if failure_reasons.empty?
-          puts "No specific issues detected - analyzing generic React PDF failure patterns..."
-
-          # This should never happen with our comprehensive checks, but just in case
-          failure_reasons << "React PDF silent failure - component rendering error not caught by try/catch block"
-          failure_reasons << "Possible @react-pdf/renderer internal bug with complex form data"
-          failure_reasons << "Memory allocation failure during PDF generation not properly reported"
-          confidence_scores["generic_react_failure"] = 50
-        end
-
-        # Sort reasons by confidence score and display
-        puts "FAILURE ANALYSIS COMPLETE - #{failure_reasons.count} potential causes identified:"
-
-        # Ensure we have confidence scores for all reasons
-        default_scores = [50] * failure_reasons.count
-        actual_scores = confidence_scores.values.presence || default_scores
-
-        # Handle mismatched array lengths (more reasons than scores)
-        if actual_scores.length < failure_reasons.length
-          puts "Warning: #{failure_reasons.length} reasons but only #{actual_scores.length} confidence scores"
-          actual_scores +=
-            [50] * (failure_reasons.length - actual_scores.length)
-        end
-
-        sorted_reasons =
-          failure_reasons
-            .zip(actual_scores)
-            .sort_by { |_, confidence| -(confidence || 50) }
-
-        sorted_reasons.each_with_index do |(reason, confidence), index|
-          confidence_text =
-            confidence >= 90 ?
-              "VERY LIKELY" :
-              confidence >= 70 ?
-                "LIKELY" :
-                confidence >= 50 ? "POSSIBLE" : "UNLIKELY"
-          puts "  #{index + 1}. [#{confidence_text}] #{reason}"
-        end
-
-        # Always provide a PRIMARY reason (highest confidence)
-        primary_reason = sorted_reasons.first[0]
-        puts ""
-        puts "🎯 PRIMARY FAILURE REASON: #{primary_reason}"
-
-        puts "Attempting Ruby fallback PDF generation..."
-
-        # Generate fallback PDF using Ruby
-        if pdf[:key] == PermitApplication::PERMIT_APP_PDF_DATA_KEY
-          generate_ruby_permit_pdf(path, submission_version)
-        elsif pdf[:key] == PermitApplication::CHECKLIST_PDF_DATA_KEY
-          generate_ruby_checklist_pdf(path, submission_version)
-        end
-
-        # Check if fallback generation worked
-        unless File.exist?(path)
-          raise "Both Node.js and Ruby fallback PDF generation failed: #{path}"
-        end
-
-        puts "Ruby fallback PDF generation successful: #{File.size(path)} bytes"
-        puts "=== End Diagnostics ==="
-      end
-
-      # Validate PDF file size
-      file_size = File.size(path)
-      if file_size < 1000 # PDFs should be at least 1KB
-        puts "Generated PDF file is too small (#{file_size} bytes), attempting Ruby fallback..."
-
-        # Remove corrupted file and generate with Ruby
-        File.delete(path) if File.exist?(path)
-
-        if pdf[:key] == PermitApplication::PERMIT_APP_PDF_DATA_KEY
-          generate_ruby_permit_pdf(path, submission_version)
-        elsif pdf[:key] == PermitApplication::CHECKLIST_PDF_DATA_KEY
-          generate_ruby_checklist_pdf(path, submission_version)
-        end
-
-        # Re-validate after Ruby generation
-        unless File.exist?(path) && File.size(path) >= 1000
-          raise "PDF generation failed - both Node.js and Ruby methods produced invalid files: #{path}"
-        end
-
-        file_size = File.size(path)
-        puts "Ruby fallback PDF generation successful after corruption: #{file_size} bytes"
-      end
-
-      puts "PDF validated successfully: #{file_size} bytes"
-
-      File.open(path) do |file|
-        doc =
-          submission_version
-            .supporting_documents
-            .where(
-              permit_application_id: submission_version.permit_application_id,
-              data_key: pdf[:key]
-            )
-            .first_or_initialize
-
-        doc.update(file:) if doc.file.blank?
-      end
-
-      # Delete the temporary file after it's been saved
-      File.delete(path) if File.exist?(path)
-    end
-
-    # Clean up JSON file after processing
-    File.delete(json_filename) if File.exist?(json_filename)
-  end
-
-  # Ruby-based PDF generation methods
-  def generate_ruby_permit_pdf(output_path, submission_version)
-    permit_application = submission_version.permit_application
-
-    content = []
-    content << "BC ENERGY STEP CODE"
-    content << "=" * 50
-    content << "PERMIT APPLICATION"
-    content << "=" * 50
-    content << ""
-    content << "Reference #: #{permit_application.number}"
-    content << "Address: #{permit_application.full_address}"
-    content << "Submission Date: #{submission_version.created_at&.strftime("%Y-%m-%d")}"
-    content << "Applicant: #{applicant_name(permit_application)}"
-    content << "Status: #{permit_application.status&.humanize}"
-    content << ""
-    content << "FORM DATA"
-    content << "=" * 30
-
-    # Add form data
-    if submission_version.formatted_submission_data.present?
-      flatten_form_data(
-        submission_version.formatted_submission_data
-      ).each do |key, value|
-        next if value.blank?
-        content << "#{key}: #{value}"
-      end
-    end
-
-    content << ""
-    content << "Generated on: #{Time.current.strftime("%Y-%m-%d %H:%M:%S")}"
-
-    # Create basic PDF structure
-    pdf_content = create_basic_pdf_structure(content.join("\n"))
-    File.write(output_path, pdf_content)
-
-    puts "Ruby permit PDF created: #{File.size(output_path)} bytes"
-  end
-
-  def generate_ruby_checklist_pdf(output_path, submission_version)
-    permit_application = submission_version.permit_application
-
-    content = []
-    content << "BC ENERGY STEP CODE"
-    content << "=" * 50
-    content << "STEP CODE CHECKLIST"
-    content << "=" * 50
-    content << ""
-    content << "Reference #: #{permit_application.number}"
-    content << "Address: #{permit_application.full_address}"
-    content << ""
-
-    if submission_version.has_step_code_checklist?
-      content << "CHECKLIST DATA"
-      content << "=" * 30
-      checklist_data = submission_version.step_code_checklist_json
-      if checklist_data.present?
-        flatten_form_data(checklist_data).each do |key, value|
-          next if value.blank?
-          content << "#{key}: #{value}"
-        end
-      end
-    else
-      content << "No step code checklist required for this application."
-    end
-
-    content << ""
-    content << "Generated on: #{Time.current.strftime("%Y-%m-%d %H:%M:%S")}"
-
-    # Create basic PDF structure
-    pdf_content = create_basic_pdf_structure(content.join("\n"))
-    File.write(output_path, pdf_content)
-
-    puts "Ruby checklist PDF created: #{File.size(output_path)} bytes"
-  end
-
-  def create_basic_pdf_structure(text_content)
-    # Create a minimal but valid PDF structure
-    clean_text = text_content.gsub(/[^\x20-\x7E\n]/, "").strip
-
-    pdf_header = "%PDF-1.4\n"
-
-    catalog = "1 0 obj\n<< /Type /Catalog /Pages 2 0 R >>\nendobj\n"
-
-    pages = "2 0 obj\n<< /Type /Pages /Kids [3 0 R] /Count 1 >>\nendobj\n"
-
-    page =
-      "3 0 obj\n<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Resources << /Font << /F1 4 0 R >> >> /Contents 5 0 R >>\nendobj\n"
-
-    font =
-      "4 0 obj\n<< /Type /Font /Subtype /Type1 /BaseFont /Courier >>\nendobj\n"
-
-    # Split text into lines and limit line length, escape parentheses
-    lines =
-      clean_text
-        .split("\n")
-        .map { |line| line[0..80].gsub(/[()\\]/, '\\\\\\&') }
-
-    content_stream = "BT\n/F1 10 Tf\n72 720 Td\n"
-    lines.each_with_index do |line, index|
-      content_stream += "0 -12 Td\n" if index > 0
-      content_stream += "(#{line}) Tj\n"
-    end
-    content_stream += "ET"
-
-    content_length = content_stream.length
-    content_obj =
-      "5 0 obj\n<< /Length #{content_length} >>\nstream\n#{content_stream}\nendstream\nendobj\n"
-
-    # Calculate correct offsets for xref table
-    objects = [pdf_header, catalog, pages, page, font, content_obj]
-    offsets = []
-    current_pos = 0
-
-    objects.each_with_index do |obj, i|
-      if i == 0
-        offsets << 0 # Header doesn't count
-      else
-        offsets << current_pos
-      end
-      current_pos += obj.length
-    end
-
-    xref_position = current_pos
-
-    xref = "xref\n0 6\n0000000000 65535 f \n"
-    (1..5).each { |i| xref += sprintf("%010d 00000 n \n", offsets[i]) }
-
-    trailer =
-      "trailer\n<< /Size 6 /Root 1 0 R >>\nstartxref\n#{xref_position}\n%%EOF"
-
-    objects.join("") + xref + trailer
-  end
-
-  def applicant_name(permit_application)
-    submitter = permit_application.submitter
-    if submitter
-      "#{submitter.first_name} #{submitter.last_name}".strip
-    else
-      "Unknown"
-    end
-  end
-
-  def flatten_form_data(data, prefix = "")
-    result = {}
-    return result unless data.is_a?(Hash)
-
-    data.each do |key, value|
-      next if key == "section-completion-key"
-
-      full_key = prefix.empty? ? key.humanize : "#{prefix} > #{key.humanize}"
-
-      case value
-      when Hash
-        result.merge!(flatten_form_data(value, full_key))
-      when Array
-        result[full_key] = value.join(", ") unless value.empty?
-      else
-        result[full_key] = value unless value.blank?
-      end
-    end
-
-    result
-  end
-
-  private
-
-  def count_nested_components(components)
-    return 0 unless components.is_a?(Array)
-
-    count = components.size
-    components.each do |component|
-      if component.is_a?(Hash) && component["components"].is_a?(Array)
-        count += count_nested_components(component["components"])
-      end
-    end
-    count
-  end
-
-  def camelize_response(data)
-    return nil if data.blank?
-    return data unless data.respond_to?(:each_pair)
-
-    camelized = {}
-    data.each_pair do |key, value|
-      camel_key = key.to_s.camelize(:lower)
-      camelized[camel_key] = case value
-      when Hash
-        camelize_response(value)
-      when Array
-        value.map { |v| v.is_a?(Hash) ? camelize_response(v) : v }
-      else
-        value
-      end
-    end
-    camelized
-  end
-end
